+++ src/_core/index.ts
@@ -2,3 +2,4 @@ export * from './Component';
 export * from './RestClient';
 export * from './Router';
 export * from './Store';
+export * from './Observer';
