+++ src/adapter/todoAdapter.ts
@@ -1,4 +1,4 @@
-import { RestClient } from "@/core";
+import { RestClient } from "@/_core";
 
 export const todoAdapterURL = 'https://js-todo-list-9ca3a.df.r.appspot.com/api';
 export const todoAdapterClient: RestClient = new RestClient(todoAdapterURL);
