+++ src/components/Team/TeamList.ts
@@ -1,4 +1,4 @@
-import {Component} from "@/core";
+import {Component} from "@/_core";
 import {SET_OPENED_TEAM_APPEND_FORM, teamStore} from "@/store";
 import {todoRouter} from "@/router";
 import {selectElement, selectParent} from "@/utils";
