+++ src/components/Todo/TodoItemAppender.ts
@@ -1,7 +1,6 @@
-import {Component} from "@/core";
+import {Component} from "@/_core";
 import {ADD_ITEM, todoOfTeamStore} from "@/store/todoOfTeamStore";
 import {KeyEvent} from "@/domains";
-import {teamStore} from "@/store";
 
 export const TodoItemAppender = class extends Component<{ id: string }> {
